refactor: Tag 컴포넌트 prop 변경 및 원할한 빌드 tsc를 위해 storybook 분리
