+++ src/stories/Badge.stories.tsx
@@ -1,6 +1,6 @@
 import { Meta, StoryObj } from '@storybook/react';
 
-import Badge from './Badge';
+import Badge from '../components/Badge/Badge';
 import { useEffect, useRef, useState } from 'react';
 
 const meta = {
