+++ src/stories/Box.stories.tsx
@@ -1,6 +1,6 @@
 import { Meta, StoryObj } from '@storybook/react';
 
-import Box from './Box';
+import Box from '../components/Box/Box';
 
 const meta = {
   title: 'Components/Box',
